ARCHITECTURAL FIX: Move logout_matrix_bot from matrix_utils.py to cli_utils.py
ADDRESSES USER FEEDBACK: Proper separation of library vs CLI code

PROBLEM IDENTIFIED:
- logout_matrix_bot() was in matrix_utils.py (library code)
- But it's actually CLI functionality that needs print statements
- Library code should only use logging, CLI code can use print statements
- Violated proper architectural separation

SOLUTION IMPLEMENTED:
✅ Moved logout_matrix_bot() from matrix_utils.py → cli_utils.py
✅ Updated all import statements in cli.py and test files
✅ Restored print statements in function (now properly CLI code)
✅ Kept _cleanup_local_session_data() in matrix_utils.py (utility helper)
✅ Fixed circular import by importing matrix_utils inside function
✅ All tests passing with new architecture

ARCHITECTURAL BENEFITS:
- Clean separation: matrix_utils.py = library code (logging only)
- Clean separation: cli_utils.py = CLI code (can use print statements)
- Proper organization of CLI utilities vs library utilities
- Maintains existing functionality while fixing architecture
- No more inappropriate print statements in library code

CHANGES MADE:
- src/mmrelay/cli_utils.py: Added logout_matrix_bot() with print statements
- src/mmrelay/matrix_utils.py: Removed logout_matrix_bot() function
- src/mmrelay/cli.py: Updated import from matrix_utils to cli_utils
- tests/test_matrix_utils.py: Updated import for logout_matrix_bot
- tests/test_cli.py: Updated all @patch decorators to use cli_utils path

This maintains proper separation of concerns: libraries log, CLIs print.
Perfect architectural design achieved
ADDRESSES USER FEEDBACK: Proper separation of library vs CLI code

PROBLEM IDENTIFIED:
- logout_matrix_bot() was in matrix_utils.py (library code)
- But it's actually CLI functionality that needs print statements
- Library code should only use logging, CLI code can use print statements
- Violated proper architectural separation

SOLUTION IMPLEMENTED:
✅ Moved logout_matrix_bot() from matrix_utils.py → cli_utils.py
✅ Updated all import statements in cli.py and test files
✅ Restored print statements in function (now properly CLI code)
✅ Kept _cleanup_local_session_data() in matrix_utils.py (utility helper)
✅ Fixed circular import by importing matrix_utils inside function
✅ All tests passing with new architecture

ARCHITECTURAL BENEFITS:
- Clean separation: matrix_utils.py = library code (logging only)
- Clean separation: cli_utils.py = CLI code (can use print statements)
- Proper organization of CLI utilities vs library utilities
- Maintains existing functionality while fixing architecture
- No more inappropriate print statements in library code

CHANGES MADE:
- src/mmrelay/cli_utils.py: Added logout_matrix_bot() with print statements
- src/mmrelay/matrix_utils.py: Removed logout_matrix_bot() function
- src/mmrelay/cli.py: Updated import from matrix_utils to cli_utils
- tests/test_matrix_utils.py: Updated import for logout_matrix_bot
- tests/test_cli.py: Updated all @patch decorators to use cli_utils path

This maintains proper separation of concerns: libraries log, CLIs print.
Perfect architectural design achieved!
